Requirement #1924: refactor import/export to make it easier to add new XML importers and exporters
1. Fixed a small bug in CommandLine class.
2. Added ability to export to multiple files (e.g. distribute a large data-set over multiple xml files, rather than putting it all into 1 file).
3. Added imex support for configuration data (ConfigurationImex class).

+++ Enterprise/Core/Imex/ImexUtils.cs
@@ -40,17 +40,47 @@ namespace ClearCanvas.Enterprise.Core.Imex
             return dataClasses.ToArray();
         }
 
-        public static void PrintImexDataClasses(TextWriter writer)
+        /// <summary>
+        /// Finds the imex that supports the specified data-class.
+        /// </summary>
+        /// <param name="dataClass"></param>
+        /// <returns></returns>
+        /// <exception cref="NotSupportedException">Indicates that no imex was found that supports the specified data-class.</exception>
+        public static IXmDataImex FindImexForDataClass(string dataClass)
         {
-            foreach (string w in ListImexDataClasses())
-                writer.WriteLine(w);
+            return (IXmDataImex)new XmlDataImexExtensionPoint().CreateExtension(
+                delegate(ExtensionInfo info)
+                {
+                    return CollectionUtils.Contains(AttributeUtils.GetAttributes<ImexDataClassAttribute>(info.ExtensionClass),
+                        delegate(ImexDataClassAttribute a)
+                        {
+                            return a != null && a.DataClass.Equals(
+                                dataClass, StringComparison.InvariantCultureIgnoreCase);
+                        });
+                });
         }
 
+        public static int GetItemsPerFile(Type imexClass)
+        {
+            ImexDataClassAttribute a = AttributeUtils.GetAttribute<ImexDataClassAttribute>(imexClass);
+            return a == null ? 0 : a.ItemsPerFile;
+        }
 
         public static void ExportToSingleFile(IXmDataImex imex, string path)
         {
+            string dir = Path.GetDirectoryName(path);
+            if (dir == path)
+                throw new ArgumentException("Path must specify a file name.");
+
             // delete the file if it exists (no error if it doesn't)
-            File.Delete(path);
+            if (File.Exists(path))
+                File.Delete(path);
+
+            // create directories as needed
+            if (!string.IsNullOrEmpty(dir) && !Directory.Exists(dir))
+            {
+                Directory.CreateDirectory(dir);
+            }
 
             using (StreamWriter writer = new StreamWriter(File.OpenWrite(path)))
             {
@@ -72,6 +102,59 @@ namespace ClearCanvas.Enterprise.Core.Imex
                 }
                 xmlWriter.WriteEndElement();
                 xmlWriter.Close();
+            }
+        }
+
+        public static void Export(IXmDataImex imex, string directory, string baseFileName, int itemsPerFile)
+        {
+            if (!Directory.Exists(directory))
+                Directory.CreateDirectory(directory);
+
+            if (itemsPerFile == 0)
+                itemsPerFile = int.MaxValue;
+
+            int itemCount = 0;
+            int fileCount = 0;
+
+            StreamWriter sw = null;
+            XmlTextWriter writer = null;
+            using (PersistenceScope scope = new PersistenceScope(PersistenceContextType.Read))
+            {
+                foreach (IExportItem item in imex.Export((IReadContext)PersistenceScope.Current))
+                {
+                    if(itemCount % itemsPerFile == 0)
+                    {
+                        // close current file
+                        if(writer != null)
+                        {
+                            writer.WriteEndElement();
+                            writer.Close();
+                            sw.Close();
+                        }
+
+                        // start new file
+                        string file = Path.Combine(directory, baseFileName + (++fileCount) + ".xml");
+
+                        // delete if already exists
+                        File.Delete(file);
+
+                        sw = new StreamWriter(File.OpenWrite(file));
+                        writer = new XmlTextWriter(sw);
+                        writer.Formatting = System.Xml.Formatting.Indented;
+                        writer.WriteStartElement(RootTag);
+                    }
+
+                    item.Write(writer);
+
+                    itemCount++;
+                }
+                scope.Complete();
+            }
+            if (writer != null)
+            {
+                writer.WriteEndElement();
+                writer.Close();
+                sw.Close();
             }
         }
